requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
facenet-pytorch==0.1.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==0.13.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.6.2
numpy==1.16.5
opencv-python==4.5.3.56
pandas==1.3.1
Pillow==6.1.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
requests==2.26.0
scikit-image==0.18.2
scikit-learn==0.21.3
scipy==1.7.0
six==1.16.0
SQLAlchemy==1.4.22
tifffile==2021.7.2
torch==1.2.0+cpu
torchvision==0.4.0+cpu
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.0
zipp==3.5.0
